+++ src/parsetypes/__init__.py
@@ -8,7 +8,7 @@
 """
 
 
-__version__ = "0.2.6"
+__version__ = "0.3"
 
 from ._common import AnyScalar, AnyScalarType, AnyValue, AnyValueType, GenericValue, Nullable
 from ._parser import TypeParser